Attempting to add gamepad support to Linux
Attempting to add gamepad support to Linux. This seems to not be working as memory becomes invalid in my implementation. I don't yet know what is going wrong so I might try to learn how to use something like Valgrind or just scrap this current implementation and try again from scratch.

Moved to a new branch to keep code as reference (since libudev was a bit confusing for me) but may replace this branch depending on how things go.