Documented List, public List struct write only now
Added documentations to `List` type and functions.

Updated Gamepad functions documentation to mention that the returned Gamepad pointer is still owned Mino's `List` type and that Mino could potentially move the value. This means that a pointer returned by WindowGetGamepad could be invalid after a call to WindowUpdate.

TODO: should I keep `WindowGetFirstConnectedGamepad`? It really only returns the first gamepad whether it is connected or not (also nil or not). Maybe change it to return the first controller in the `List` that is connected?

Modified `List` struct to be write only for public api. This is probably unnecessary but `List` isn't meant to be modified outside of Mino. This just deters users from modifying the values and inadvertently putting the `List` in an invalid state.