OpenGL on Linux
Mino can now set up OpenGL on Linux with X11. the graphics API is kinda changing to be bit bare bones though. Instead of attempting to abstract the OpenGL function calls into draw functions in Mino, It's much simpler to just provide functions to create the OpenGL context, make it current, and then discard the context. The rest is mostly crossplatform (Linux and Windows both have `GL/gl.h` files which you can use to interact with Mino's GL context).

This change means that I also need to fix this on the windows platform.