Quantum Leap branch
I don't have much more time for the game jam this library is for so changes are being seperated from `main` for a bit.

Noticed that the amalgamation files being in `./project` might cause problems with this make file so removed them for now. maybe they could just be printed to stdout and users can write the output to their own files?

Also noticed that the Sleep function in Windows is pretty inaccurate so created a more precise sleep function (at least on Windows).