player Rotation, bounding, and asteroids
Player is bounded to window and rotates to the direction if it's acceleration. Also started creating asteroids.

I won't be able to finish the game in time for the jam but I might still just work on polishing Mino for next time (or to just use in my own projects).