New CMake and plans for major restructuring
Finally begun to understand CMake a little bit and have started to convert Mino to use CMake instead of plain Makefiles.

This is mostly intended for preperations to add support for macOS and other platforms such as WebAssembly. \

Mino was originally meant to be "Pure C". However, most of Apple's non-deprecated libraries appear to be meant for Objective-C or Swift. Because of this, Mino will start to use Objective-C for macOS (and to be fair to Windows and Linux, Mino will also use some C++) but this planned to be strictly for platform specific sources (the user facing API will stay plain C but the interface between Mino and the OS will start to use other languages).

Also because of these planned changes and familiarizing myself with CMake, I've started to move platform specific sources into their own directories for better organization (and to keep files small and easily navigable).

Lastly I feel like while I am at it, I should change up Mino's API a little bit to be more flexible. Thinking about things like multi-window support, trying to have options to be less reliant on the C Standard Library, better graphic API support (with more backends than just OpenGL), and try to see where I can add the some of the (currently private) datastructure types I had been experimenting with (of course try to make it optional because size is still a priority to Mino).